requirements.txt
streamlit==1.31.1
openai==1.16.0
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.13
langchain-pinecone==0.2.6
python-dotenv==1.0.0
pandas==2.2.0
altair==5.2.0
matplotlib==3.8.2
numpy==1.26.3

# Vector database
pinecone-client==2.2.4

# LangSmith for tracing
langsmith==0.1.80

# Evaluation and fine-tuning
ragas==0.1.0
datasets==2.16.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
